Add intermediate task for transformation of model information
- This prevents issues when assigning ids and processing links in the same task
