+++ etl/jobs/transformation/model_transformer_job.py
@@ -12,29 +12,31 @@ from etl.jobs.util.id_assigner import add_id
 
 def main(argv):
     """
-    Creates a parquet file with model data.
+    
     :param list argv: the list elements should be:
-                    [1]: Parquet file path with raw model data
-                    [2]: Parquet file path with publication group data
-                    [3]: Output file
+                    [1]: Parquet file path with raw external_model_ids configuration
+                    [2]: Parquet file path with initial df with model_information data
+                    [3]: Parquet file path with publication group data
+                    [4]: Parquet file path with accessibility group data
+                    [5]: Parquet file path with contact people data
+                    [6]: Parquet file path with contact form data
+                    [7]: Parquet file path with source database data
+                    [8]: Parquet file path with license data
+                    [9]: Output file
     """
-    raw_model_parquet_path = argv[1]
-    raw_cell_model_parquet_path = argv[2]
-    raw_sharing_parquet_path = argv[3]
-    raw_external_model_ids_resources_parquet_path = argv[4]
-    publication_group_parquet_path = argv[5]
-    accessibility_group_parquet_path = argv[6]
-    contact_people_parquet_path = argv[7]
-    contact_form_parquet_path = argv[8]
-    source_database_parquet_path = argv[9]
-    license_parquet_path = argv[10]
-    output_path = argv[11]
+    raw_external_model_ids_resources_parquet_path = argv[1]
+    initial_model_information_parquet_path = argv[2]
+    publication_group_parquet_path = argv[3]
+    accessibility_group_parquet_path = argv[4]
+    contact_people_parquet_path = argv[5]
+    contact_form_parquet_path = argv[6]
+    source_database_parquet_path = argv[7]
+    license_parquet_path = argv[8]
+    output_path = argv[9]
 
     spark = SparkSession.builder.getOrCreate()
-    raw_model_df = spark.read.parquet(raw_model_parquet_path)
-    raw_cell_model_df = spark.read.parquet(raw_cell_model_parquet_path)
-    raw_sharing_df = spark.read.parquet(raw_sharing_parquet_path)
     raw_external_model_ids_df = spark.read.parquet(raw_external_model_ids_resources_parquet_path)
+    initial_model_information_df = spark.read.parquet(initial_model_information_parquet_path)
     publication_group_df = spark.read.parquet(publication_group_parquet_path)
     accessibility_group_df = spark.read.parquet(accessibility_group_parquet_path)
     contact_people_df = spark.read.parquet(contact_people_parquet_path)
@@ -43,10 +45,8 @@ def main(argv):
     license_df = spark.read.parquet(license_parquet_path)
 
     model_df = transform_model(
-        raw_model_df,
-        raw_cell_model_df,
-        raw_sharing_df,
         raw_external_model_ids_df,
+        initial_model_information_df,
         publication_group_df,
         accessibility_group_df,
         contact_people_df,
@@ -58,10 +58,8 @@ def main(argv):
 
 
 def transform_model(
-        raw_model_df: DataFrame,
-        raw_cell_model_df: DataFrame,
-        raw_sharing_df: DataFrame,
         raw_external_model_ids_df: DataFrame,
+        initial_model_information_df: DataFrame,
         publication_group_df: DataFrame,
         accessibility_group_df: DataFrame,
         contact_people_df: DataFrame,
@@ -69,9 +67,7 @@ def transform_model(
         source_database_df: DataFrame,
         license_df: DataFrame) -> DataFrame:
 
-    model_df = get_data_from_model_modules(raw_model_df, raw_cell_model_df)
-    model_df = join_model_with_sharing(model_df, raw_sharing_df)
-    model_df = add_id(model_df, "id")
+    model_df = initial_model_information_df
     model_df = set_fk_publication_group(model_df, publication_group_df)
     model_df = set_fk_accessibility_group(model_df, accessibility_group_df)
     model_df = set_fk_contact_people(model_df, contact_people_df)
@@ -82,85 +78,6 @@ def transform_model(
     
     model_df = get_columns_expected_order(model_df)
 
-    return model_df
-
-
-def get_data_from_model_modules(raw_model_df: DataFrame, raw_cell_model_df: DataFrame) -> DataFrame:
-    model_df = raw_model_df.select(
-        "model_id", 
-        "publications", 
-        "external_ids",
-        "supplier",
-        "supplier_type",
-        "catalog_number",
-        "vendor_link",
-        lit("").alias("rrid"),
-        "parent_id",
-        "origin_patient_sample_id",
-        lit("").alias("model_name"),
-        lit("").alias("model_name_aliases"),
-        lit("").alias("growth_properties"),
-        lit("").alias("growth_media"),
-        lit("").alias("media_id"),
-        lit("").alias("plate_coating"),
-        lit("").alias("other_plate_coating"),
-        lit("").alias("passage_number"),
-        lit("").alias("contaminated"),
-        lit("").alias("contamination_details"),
-        lit("").alias("supplements"),
-        lit("").alias("drug"),
-        lit("").alias("drug_concentration"),
-        Constants.DATA_SOURCE_COLUMN).drop_duplicates()
-    model_df = model_df.withColumn("type", lit("PDX"))
-    model_df = model_df.withColumnRenamed("model_id", "external_model_id")
-
-    cell_model_df = raw_cell_model_df.select(
-        "model_id",
-        "publications",
-        "external_ids",
-        "supplier",
-        "supplier_type",
-        "catalog_number",
-        "vendor_link",
-        "rrid",
-        "parent_id",
-        "origin_patient_sample_id",
-        "type",
-        "model_name",
-        "model_name_aliases",
-        "growth_properties",
-        "growth_media",
-        "media_id",
-        "plate_coating",
-        "other_plate_coating",
-        "passage_number",
-        "contaminated",
-        "contamination_details",
-        "supplements",
-        "drug",
-        "drug_concentration",
-        Constants.DATA_SOURCE_COLUMN).drop_duplicates()
-    cell_model_df = cell_model_df.withColumn("type", lower_and_trim_all("type"))
-    cell_model_df = cell_model_df.withColumnRenamed("model_id", "external_model_id")
-    
-    # Standardise some of the model type values
-    cell_model_df = cell_model_df.withColumn(
-        "type",
-        when((col("type") == 'cell line'), "cell line")
-        .when((col("type").like('%organoid%')), "organoid")
-        .otherwise(lit("other"))
-    )
-
-    union_df = model_df.unionByName(cell_model_df)
-
-    return union_df
-
-
-def join_model_with_sharing(model_df: DataFrame, raw_sharing_df: DataFrame) -> DataFrame:
-    raw_sharing_df = raw_sharing_df.withColumnRenamed("model_id", "external_model_id")
-    model_df = model_df.join(
-        raw_sharing_df,
-        on=['external_model_id', Constants.DATA_SOURCE_COLUMN], how='left')
     return model_df
