+++ luigi_template.cfg
@@ -87,6 +87,7 @@ log_level = INFO
 [ExtractModelIdsResources]
 
 [TransformPatient]
+[TransformInitialModel]
 [TransformModel]
 [TransformLicense]
 [TransformCellModel]
